Web_App/requirements.txt
click==7.1.1
cycler==0.10.0
enum34==1.1.10
Flask==1.1.2
itsdangerous
Jinja2
kiwisolver
MarkupSafe
matplotlib
numpy
opencv-python
pandas
piexif
Pillow
pyparsing
python-dateutil
pytz
PyYAML
six
Werkzeug
xlrd
xlutils
xlwt
efficientnet_pytorch